move init stuff into init file
